fixing setup and adding channel

+++ python/test/test_txrequest.py
@@ -1,10 +1,10 @@
-from pytwisterx.api.comms.channel import TxRequest
+from pytwisterx.api.comms.request import TxRequest
 import numpy as np
-buf = np.array([1,2,3,4,5,6,7,8], dtype=np.int32)
+buf = np.array([1,2,3,4,5,6,7,8], dtype=np.double)
 tx = TxRequest(10, buf, 8, np.array([1,2,3,4], dtype=np.int32), 4)
 
 print(tx.target, tx.buf, tx.header, tx.headerLength, tx.length)
 
 print(type(tx.buf), type(tx.header))
 
-print(tx.to_string(b'int', 32))
+print(tx.to_string(b'double', 32))
